Add files via upload
This is the same multi-linear regression model as what I posted previously with a few adjustments. I added a separate dataframe that generates macro data along with stock data for the same months (rather than lagging the macro data). This will allow you to apply the coefficients to the macro data to project stocks/etfs. this model feeds into this tableau report as well:
https://public.tableau.com/app/profile/adam.d4129/viz/MultipleLinearRegressionPredictiveModelforSPY/1YearLeadingIndicatorsComparedtoSPYMultilinearRegression