Create Stock Buy Sell Pricing
This code pulls all S&P 500 stock data for the specified time frame, calculates various functions, and aims at refining the list to stocks with low slope, and high variance and identifying which are likely over valued or undervalued in the short term based on random market fluctuations and not based on fundamental values. This is by no means investment recommendations, just an example of my continuing journey for learning Python and applying it to investments.