Stock Trading Algo Test
This code identifies all stocks currently trading in the S&P 500, then it determines which are currently trading above or below 1 standard deviation from their average price in the specified amount of days. It then refines the list to only show stocks with low slope, and high variance. This is by no means an investment recommendation, just a way for me to practice python and experiment with investment data.