Add files via upload
This program pulls all stocks in the S&P 500 and determines if they are low slope, and high volatility. It also determines if they are above or below their standard deviation for the time specified. This is by no means an investment recommendation, it is just a fun project for me to practice Python.