Add files via upload
This is an extension to my stock screening algo file, This file will not only screen all stocks on the S&P for over value and under value, it will also place hypothetical trades in a hypothetical account and document the history of trading. Again, this is not an investment recommendation, this is just a fun project for me to practice Python.