Add files via upload
This is my Stock Screening Algo with a trading algo attached. It places buys/sells in the stocks that are identified as under or over valued. It also keeps an ongoing reccords and tracks current hypothetical holding. Again, this is not an investment recommendation, it is just a way for me to practice Python.