Add files via upload
This is my stock screening algorithm, which identifies stocks in the S&P 500 that are low slope, and highly volatile, and includes an attached trading algorithm. The program places hypothetical trades and records them in an ongoing ledger. It also will post hypothetical holdings. Again, this is not an investment recommendation, this is just a way for me to practice python and writing algorithm.